修复oracle下的运行问题；完善代码生成器页面功能；移除api模块的代码生成器，统一使用web模块的生成器；修改t_s_column_ui的id列为long类型；SQL增量更新；element 2.8.2存在表格组件异步加载树形数据无法撑开的bug，暂回退到2.7.2版本。

+++ web/src/main/java/com/simon/service/ColumnUiService.java
@@ -9,7 +9,7 @@ import java.util.List;
 * @author SimonSun
 * @date 2019-05-03
 **/
-public interface ColumnUiService extends BasicService<ColumnUi, Integer> {
+public interface ColumnUiService extends BasicService<ColumnUi, Long> {
     /**
      * 根据表名查询
      * @param tableName 表名
